pyproject.toml
[tool.poetry]
name = "smard"
version = "0.1.0"
description = ""
authors = ["Stefan Hudelmaier <stefan@huditech.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
paho-mqtt = "2.1.0"
python-dotenv = "1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
